request body and parms validation

+++ src/app/controllers/ratingController.js
@@ -3,8 +3,9 @@ const apiResponse = require("../helpers/apiResponse");
 const logger = require("../utils/logger");
 const fetchCustomerApi = require("../domain/service/fetchCustomerApi");
 const fetchRestaurantApi = require("../domain/service/fetchRestaurantApi");
-const amqp_connection_string = "amqps://vpcafnvn:IZKTpMMbAzj-anWfqkEWM0N47tWXvxKo@elk.rmq2.cloudamqp.com/vpcafnvn";
+const amqp_connection_string = process.env.amqp_connection_string;
 const amqp_task_queue_str = "rating_post";
+const {  validationResult } = require('express-validator');
 
 exports.getAllRating = async (req, res) => {
     try {
@@ -21,18 +22,27 @@ exports.getAllRating = async (req, res) => {
     }
 }
 
-exports.getSingleRating = async (req, res) => {
+exports.getSingleRating = async (req, res, next) => {
     try {
+        const validationErrors = validationResult(req);
+        if (!validationErrors.isEmpty()) {
+            return res.status(400).json({ errors: validationErrors.array() });
+        }
         const restaurant = await fetchRestaurantApi.getRestaurant(req.params.restaurantId, req.user.token);
-        if (restaurant) {
+        if (restaurant.status) {
             const rating = await ratingService.getSingleRating(req.params.rateId, req.user, restaurant);
             return apiResponse.successResponseWithData(res, "Single rating record found", rating);
         } else {
-            throw new Error("No such restaurant found");
+            const error = new Error("No such restaurant found");
+            error.statusCode = 404;
+            throw error;
         }
     } catch (error) {
-        logger.error(error.message);
-        return apiResponse.customErrorResponse(res, error.message, 404);
+        console.log("getSingleRating ======>", error)
+        // if (!error.statusCode) {
+        error.statusCode = 404;
+        // }
+        next(error);
     }
 }
 
@@ -44,20 +54,20 @@ exports.createRating = async (req, res) => {
         if (restaurant) {
             body.restaurantId = req.params.restaurantId;
             const rating = await ratingService.createRating(body);
-            let open =  require('amqplib').connect(amqp_connection_string);
-            open.then(function(conn) {
+            let open = require('amqplib').connect(amqp_connection_string);
+            open.then(function (conn) {
                 return conn.createChannel();
-              }).then(function(ch) {
-                return ch.assertQueue(amqp_task_queue_str).then(function(ok) {
-                  return ch.sendToQueue(amqp_task_queue_str, Buffer.from(JSON.stringify({userRating:rating})));
+            }).then(function (ch) {
+                return ch.assertQueue(amqp_task_queue_str).then(function (ok) {
+                    return ch.sendToQueue(amqp_task_queue_str, Buffer.from(JSON.stringify({ userRating: rating })));
                 });
-              }).catch(err =>console.log("error amqp=======>",err));
+            }).catch(err => console.log("error amqp=======>", err));
             return apiResponse.successResponseWithData(res, "Rating review add successfully", rating);
         } else {
             throw new Error("No such restaurant found");
         }
     } catch (error) {
-        console.log("error ====>",error);
+        console.log("error ====>", error);
         logger.error(error.message);
         return apiResponse.customErrorResponse(res, error.message, 406);
     }
@@ -65,6 +75,10 @@ exports.createRating = async (req, res) => {
 
 exports.updateRating = async (req, res) => {
     try {
+        const validationErrors = validationResult(req);
+        if (!validationErrors.isEmpty()) {
+            return res.status(400).json({ errors: validationErrors.array() });
+        }
         let body = req.body;
         const updateRating = await ratingService.updateRating(body, req.params.rateId);
         return apiResponse.successResponseWithData(res, "Rating review updated successfully", updateRating);
@@ -77,6 +91,10 @@ exports.updateRating = async (req, res) => {
 
 exports.updateRatingStatus = async (req, res) => {
     try {
+        const validationErrors = validationResult(req);
+        if (!validationErrors.isEmpty()) {
+            return res.status(400).json({ errors: validationErrors.array() });
+        }
         let body = req.body;
         const updateRatingStatus = await ratingService.updateRatingStatus(body, req.params.rateId);
         return apiResponse.successResponseWithData(res, "Rating review status updated successfully", updateRatingStatus);
@@ -89,6 +107,10 @@ exports.updateRatingStatus = async (req, res) => {
 
 exports.deleteRating = async (req, res) => {
     try {
+        const validationErrors = validationResult(req);
+        if (!validationErrors.isEmpty()) {
+            return res.status(400).json({ errors: validationErrors.array() });
+        }
         const deleteRating = await ratingService.deleteRating(rateId);
         if (deleteRating) {
             return apiResponse.successResponseWithData(res, "Rating review deleted successfully", updateRatingStatus);
